[fix] 修复任意文件删除

+++ apps/system/api/upload.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"errors"
 	"fmt"
 	"net/http"
 	"os"
@@ -11,6 +12,8 @@ import (
 	"pandax/pkg/global"
 	"pandax/pkg/tool"
 	"path"
+	"path/filepath"
+	"strings"
 	"time"
 )
 
@@ -52,6 +55,13 @@ func (up *UploadApi) GetImage(rc *restfulx.ReqCtx) {
 func (up *UploadApi) DeleteImage(rc *restfulx.ReqCtx) {
 	fileName := restfulx.QueryParam(rc, "fileName")
 	biz.NotEmpty(fileName, "请传要删除的图片名")
+
+	absFilePath := filepath.Join(filePath, fileName)
+	if !strings.HasPrefix(absFilePath, filepath.Clean(filePath)+string(filepath.Separator)) {
+		biz.ErrIsNil(errors.New("非法文件路径"), "非法文件路径")
+		return
+	}
+
 	err := os.Remove(fmt.Sprintf("%s/%s", filePath, fileName))
 	biz.ErrIsNil(err, "文件删除失败")
 }
